# Add tablet support