--- a/osu.Framework.Benchmarks/BenchmarkTabletDriver.cs
+++ b/osu.Framework.Benchmarks/BenchmarkTabletDriver.cs
@@ -1,6 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+#if NET5_0
 using BenchmarkDotNet.Attributes;
 using osu.Framework.Input.Handlers.Tablet;
 
@@ -22,3 +23,4 @@ namespace osu.Framework.Benchmarks
         }
     }
 }
+#endif