--- a/osu.Framework.Tests/Visual/Input/TestSceneTabletInput.cs
+++ b/osu.Framework.Tests/Visual/Input/TestSceneTabletInput.cs
@@ -1,12 +1,16 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Linq;
+using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input;
 using osu.Framework.Input.Events;
+using osu.Framework.Input.Handlers.Tablet;
+using osu.Framework.Platform;
 using osuTK;
 using osuTK.Graphics;
 
@@ -40,6 +44,19 @@ namespace osu.Framework.Tests.Visual.Input
                 Direction = FillDirection.Vertical,
                 Children = new[] { penButtonFlow, auxButtonFlow }
             };
+        }
+
+        [Resolved]
+        private GameHost host { get; set; }
+
+        protected override void LoadComplete()
+        {
+            base.LoadComplete();
+
+            var tabletHandler = host.AvailableInputHandlers.OfType<OpenTabletDriverHandler>().FirstOrDefault();
+
+            if (tabletHandler != null)
+                AddToggleStep("toggle tablet handling", t => tabletHandler.Enabled.Value = t);
         }
 
         private class PenButtonHandler : CompositeDrawable