--- a/osu.Framework/Input/Events/TabletAuxiliaryButtonPressEvent.cs
+++ b/osu.Framework/Input/Events/TabletAuxiliaryButtonPressEvent.cs
@@ -1,0 +1,18 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Input.States;
+
+namespace osu.Framework.Input.Events
+{
+    /// <summary>
+    /// An event representing the press of a tablet auxiliary button.
+    /// </summary>
+    public class TabletAuxiliaryButtonPressEvent : TabletAuxiliaryButtonEvent
+    {
+        public TabletAuxiliaryButtonPressEvent(InputState state, TabletAuxiliaryButton button)
+            : base(state, button)
+        {
+        }
+    }
+}