… osu.Framework/Input/Handlers/Tablet/OpenTabletDriverHandler.cs
… osu.Framework/Input/Handlers/Tablet/RelativeTabletMode.cs
… osu.Framework/Input/Handlers/Tablet/TabletDriver.cs
… osu.Framework/Input/InputManager.cs
… osu.Framework/Input/PassThroughInputManager.cs
… osu.Framework/Input/StateChanges/TabletAuxiliaryButtonInput.cs
… osu.Framework/Input/StateChanges/TabletPenButtonInput.cs
… osu.Framework/Input/States/InputState.cs
… osu.Framework/Input/States/TabletState.cs
… osu.Framework/Input/TabletAuxiliaryButton.cs
… osu.Framework/Input/TabletAuxiliaryButtonEventManager.cs
… osu.Framework/Input/TabletPenButton.cs
… osu.Framework/Input/TabletPenButtonEventManager.cs
… osu.Framework/Platform/DesktopGameHost.cs
… osu.Framework/Statistics/FrameStatistics.cs
… osu.Framework/Testing/Input/ManualInputManager.cs
… osu.Framework/Threading/InputThread.cs
… osu.Framework/osu.Framework.csproj