- Added OpenTabletDriver package with implementation for raw input

- [x] Depends on #4166