@InfinityGhost 2021-01-19
Looks like its building for .NET Core 3.1 in CI, is this updated on master?