@bdach 2021-01-19
Where are you drawing that conclusion from? inspectcode is not the compiler, fwiw.