@InfinityGhost 2021-01-19
> Where are you drawing that conclusion from? inspectcode is not the compiler, fwiw.

Nvm, made that guess based on this at first glance
`Running on AMD 64 in 64-bit mode, .NET runtime 3.1.11`