@InfinityGhost 2021-01-21
Current issue, window size change doesn't update the output area, and there doesn't appear to be any bindable or event to hook to update this.