@bdach 2021-01-21
Some pieces are already in place to get one in for SDL2 at least:

https://github.com/ppy/osu-framework/blob/0a8294590f2acb898096668591577d76367720ac/osu.Framework/Platform/SDL2DesktopWindow.cs#L448-L466

but they're not exposed in a nice enough way to be usable just yet.