@InfinityGhost 2021-01-22
> Some pieces are already in place to get one in for SDL2 at least:
> 
> https://github.com/ppy/osu-framework/blob/0a8294590f2acb898096668591577d76367720ac/osu.Framework/Platform/SDL2DesktopWindow.cs#L448-L466
> 
> but they're not exposed in a nice enough way to be usable just yet.

Should that change be in a separate PR or on this branch? I don't want to break too far out of scope of this pull request unless its fine to do so with a relatively smaller change like that.