@peppy 2021-01-26
This whole API feels very haphazard to me. Setting the `OutputMode` to an instance which is then used to actually interface with the API isn't something I've seen done before. Usually you'd be changing the mode on the library using a `Configuration` or `Enum` object and then consuming it.

Is there a specific reason for doing things the way you are? Or is that just how it ended up?