@peppy 2021-02-01
I've done an initial clean-up pass, and things look pretty good apart from the pointed out issues.

I haven't tested this yet; need to dig out my test tablets (will do so this week).