@InfinityGhost 2021-02-19
Added a benchmark for testing detection performance, here's my run.

## Results
``` ini

BenchmarkDotNet=v0.12.1, OS=nixos 21.03pre269343.758b29b5a28
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100
  [Host]     : .NET Core 5.0.0 (CoreCLR 5.0.20.51904, CoreFX 5.0.20.51904), X64 RyuJIT
  DefaultJob : .NET Core 5.0.0 (CoreCLR 5.0.20.51904, CoreFX 5.0.20.51904), X64 RyuJIT


```

#### Without a tablet plugged in
|          Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| DetectBenchmark | 604.6 ns | 6.56 ns | 5.82 ns | 0.1297 |     - |     - |     544 B |


#### With a Wacom CTL-470 plugged in
|          Method |     Mean |    Error |   StdDev |   Median | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |---------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| DetectBenchmark | 340.6 ms | 21.90 ms | 64.58 ms | 299.2 ms |     - |     - |     - | 806.61 KB |