@peppy 2021-03-05
The `TabletDriver` constructor is taking 500ms here. This is unacceptably high.