@peppy 2021-03-05
Here's a 10 second profiling effort:

![20210305 133231 (JetBrains Rider)](https://user-images.githubusercontent.com/191335/110067519-411f7580-7db7-11eb-8f69-403a52e1521a.png)

If you'd prefer I step in and try and fix the performance issues at OTD's end, let me know. This is a relatively high priority for us as we move away from the weirdly broken raw input implementation we've been relying on, so I don't mind helping get things in a working state if that works better for you.