@InfinityGhost 2021-03-06
> The `TabletDriver` constructor is taking 500ms here (with no devices connected). This is unacceptably high.

This is easily fixable, it's leftovers from the desktop plugin system that uses reflection which was really slow but versatile. https://github.com/OpenTabletDriver/OpenTabletDriver/pull/825 should resolve this, try with these changes.

Would test myself, but I'm waiting on my OSS rider license lol