@InfinityGhost 2021-03-08
Performance fixes are within the v0.5.2 update, this should be basically ready now.