@bdach 2021-03-09
On a multi-monitor linux setup I'm seeing the game receive global coordinates - as in across all desktops, so for example if I have my tablet mapped to the right display, I can only ever see the cursor if the tablet pen is at the left edge of the input (where the screens join). Can someone verify if this is only a linux thing, or will this be the case on all platforms?