@bdach 2021-03-09
So... the tablet detection refused to work for me on linux, and I traced it back to [this FAQ entry](https://github.com/OpenTabletDriver/OpenTabletDriver/wiki/Linux-FAQ#the-driver-fails-to-open-the-tablet-deviceioexception). And I have *several* questions...

Is it required for me to have to clone a separate repo and do admin-level tasks to have this work? I mean sure, it's linux, I'm fine with tinkering, but I don't really think we want to be getting user complaints about this. It should at least present a reasonable error that links to this FAQ, I think? Not to mention that the entire notion of a "usermode driver" kinda goes out the window if I have to do this... I haven't even checked what the rules do yet.