@InfinityGhost 2021-03-10
> 0666 seems egregious. rw-rw-rw-? Really? That sounds like an outrageous security hole. The impact of this should be explained. It should NOT just be a copy-paste snippet.

If you take into account this only applies to the tablet endpoints where no sensitive data would ever be passed through anyway, there's really no security hole by doing this. There's absolutely no other way to make these endpoints read/write from user mode either than changing the group as well, but in my experience its even more work to make sure that the user is in the right group too (we fought with this for like 3 releases).

> Is it required for me to have to clone a separate repo and do admin-level tasks to have this work? I mean sure, it's linux, I'm fine with tinkering, but I don't really think we want to be getting user complaints about this. It should at least present a reasonable error that links to this FAQ, I think?

In terms of making it so this is usable out of the box, udev rules can be included in AppImages as well (where you need root to install anyway) but there's no way around udev rules even with user mode drivers. Windows has the equivalent of this permission level by default, so I don't see the issue.